Move types to types package
