+++ types/index.ts
@@ -1,1 +1,4 @@
 export * from './chat'
+export * from './auth'
+export * from './lsp'
+export * from './inlineCompletionWithReferences'
