+++ types/lsp.ts
@@ -1,0 +1,3 @@
+export { TextDocument, Position } from 'vscode-languageserver-textdocument'
+
+export * from 'vscode-languageserver-types'
